requirements.txt
absl-py==2.0.0
aiosignal==1.3.1
alabaster==0.7.13
# antlr4-python3-runtime==4.8
appdirs==1.4.4
argcomplete==1.12.3
asttokens==2.4.1
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
black==20.8b1
blinker==1.7.0
Brotli==1.0.9
cachetools==4.2.4
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==2.0.4
click==8.1.7
cloudpickle==1.6.0
colorlog==4.8.0
cryptography==41.0.3
cycler==0.12.1
Cython==0.29.36
decorator==5.1.1
distlib==0.3.7
docutils==0.20.1
exceptiongroup==1.2.0
execnet==2.0.2
executing==2.0.1
fasteners==0.19
filelock==3.13.1
flake8==3.8.4
flake8-bugbear==20.11.1
flake8-comprehensions==3.3.1
flake8-docstrings==1.5.0
Flask==3.0.0
Flask-Compress==1.14
frozenlist==1.4.0
future==0.18.3
glfw==2.7.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.59.3
# gym==0.18.0
gym-notices==0.0.8
h5py==3.10.0
hiplot==0.1.22
hydra-core==1.0.5
hydra-submitit-launcher==1.0.1
identify==2.5.32
idna==3.4
imageio==2.9.0
imageio-ffmpeg==0.4.3
imagesize==1.4.1
importlib-metadata==6.8.0
incremental==22.10.0
iniconfig==2.0.0
ipython==8.17.2
isort==5.7.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
kiwisolver==1.4.5
latexcodec==2.0.1
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.3.3
matplotlib-inline==0.1.6
mccabe==0.6.1
# metaworld==0.1.0
mllogger==0.7
msgpack==1.0.7
mujoco==2.1.5
mujoco-py==2.1.2.14
multipledispatch==1.0.0
mypy==0.790
mypy-extensions==0.4.4
nodeenv==1.8.0
nox==2020.12.31
numpy==1.26.2
nvidia-ml-py==12.535.133
nvitop==1.3.1
oauthlib==3.2.2
omegaconf==2.0.6
opencv-python==4.5.1.48
packaging==23.2
pandas==1.2.0
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
Pillow==7.1.0
pip==23.3.1
platformdirs==4.0.0
pluggy==0.13.1
pockets==0.9.1
pre-commit==2.9.3
prompt-toolkit==3.0.41
protobuf==4.25.1
psutil==5.8.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybtex==0.24.0
pybtex-docutils==1.0.3
pycodestyle==2.6.0
pycparser==2.21
pydocstyle==6.3.0
pyflakes==2.2.0
pyglet==1.5.0
Pygments==2.17.2
pymongo==3.11.2
PyOpenGL==3.1.7
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyrr==0.10.3
PySocks==1.7.1
pytest==6.2.1
pytest-forked==1.6.0
pytest-xdist==2.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
ray==2.8.0
referencing==0.31.0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.13.1
rsa==4.9
scipy==1.11.4
setuptools==65.5.0
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==3.4.3
sphinx-autodoc-annotation==1.0.post1
sphinx-copybutton==0.3.1
sphinx-rtd-theme==0.5.1
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-bibtex==2.1.3
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-napoleon==0.7
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.6.3
submitit==1.5.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
toml==0.10.2
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
towncrier==19.2.0
traitlets==5.13.0
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.26.18
virtualenv==20.24.7
wcwidth==0.2.12
Werkzeug==3.0.1
wheel==0.38.0
yamllint==1.25.0
zipp==3.17.0
